Refactors Snowtooth Graph Endpoint
