+++ package.json
@@ -43,7 +43,7 @@
     "cookie-parser": "^1.4.3",
     "cors": "^2.8.1",
     "express": "^4.14.0",
-    "express-graphql": "^0.6.6",
+    "express-graphql": "^0.6.11",
     "gmail-send": "^1.2.6",
     "graphql": "^0.10.3",
     "graphql-tools": "^1.2.3",
