--- test/server/graphql/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<script src="bundle.js">
-	</script>
-	<title>Working with Babel</title>
-</head>
-<body>
-
-</body>	
-</html>
