--- test/server/graphql/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "testing-webpack",
-  "version": "1.0.0",
-  "description": "",
-  "main": "script.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [
-    "webpack",
-    "es6"
-  ],
-  "author": "Eve Porcello",
-  "license": "MIT"
-}
